requirements.txt
Django==1.11.2
djangorestframework==3.6.3
Markdown==2.6.8
MySQL-python==1.2.5
pytz==2017.2
